+++ apps/antalmanac/src/components/RightPane/CoursePane/CourseRenderPane.tsx
@@ -1,5 +1,4 @@
-import { Close } from '@mui/icons-material';
-import { Alert, Box, IconButton, useMediaQuery } from '@mui/material';
+import { Box } from '@mui/material';
 import {
     AACourse,
     AASection,
@@ -13,24 +12,17 @@ import { useCallback, useEffect, useState } from 'react';
 import LazyLoad from 'react-lazyload';
 
 import RightPaneStore from '../RightPaneStore';
-import GeDataFetchProvider from '../SectionTable/GEDataFetchProvider';
-import SectionTableLazyWrapper from '../SectionTable/SectionTableLazyWrapper';
-
-import SchoolDeptCard from './SchoolDeptCard';
-import darkModeLoadingGif from './SearchForm/Gifs/dark-loading.gif';
-import loadingGif from './SearchForm/Gifs/loading.gif';
-import darkNoNothing from './static/dark-no_results.png';
-import noNothing from './static/no_results.png';
 
 import { openSnackbar } from '$actions/AppStoreActions';
-import analyticsEnum from '$lib/analytics';
+import { ErrorMessage } from '$components/RightPane/CoursePane/Messages/ErrorMessage';
+import { LoadingMessage } from '$components/RightPane/CoursePane/Messages/LoadingMessage';
+import { RecruitmentBanner } from '$components/RightPane/CoursePane/RecruitmentBanner';
+import { SectionTableWrapped } from '$components/RightPane/SectionTable/SectionTableWrapped';
 import { Grades } from '$lib/grades';
 import { Larc } from '$lib/larc';
-import { getLocalStorageRecruitmentDismissalTime, setLocalStorageRecruitmentDismissalTime } from '$lib/localStorage';
 import { WebSOC } from '$lib/websoc';
 import AppStore from '$stores/AppStore';
 import { useHoveredStore } from '$stores/HoveredStore';
-import { useThemeStore } from '$stores/SettingsStore';
 
 function getColors() {
     const currentCourses = AppStore.schedule.getCurrentCourses();
@@ -64,131 +56,6 @@ const flattenSOCObject = (SOCObject: WebsocAPIResponse): (WebsocSchool | WebsocD
 
         return accumulator;
     }, []);
-};
-const RecruitmentBanner = () => {
-    const [bannerVisibility, setBannerVisibility] = useState(true);
-
-    const isDark = useThemeStore((store) => store.isDark);
-
-    // Display recruitment banner if more than 11 weeks (in ms) has passed since last dismissal
-    const recruitmentDismissalTime = getLocalStorageRecruitmentDismissalTime();
-    const dismissedRecently =
-        recruitmentDismissalTime !== null &&
-        Date.now() - parseInt(recruitmentDismissalTime) < 11 * 7 * 24 * 3600 * 1000;
-    const isSearchCS = ['COMPSCI', 'IN4MATX', 'I&C SCI', 'STATS'].includes(
-        RightPaneStore.getFormData().deptValue.toUpperCase()
-    );
-    const displayRecruitmentBanner = bannerVisibility && !dismissedRecently && isSearchCS;
-
-    const isMobileScreen = useMediaQuery('(max-width: 750px)');
-
-    return (
-        <Box sx={{ position: 'fixed', bottom: 5, right: isMobileScreen ? 5 : 75, zIndex: 999 }}>
-            {displayRecruitmentBanner ? (
-                <Alert
-                    icon={false}
-                    severity="info"
-                    style={{
-                        color: isDark ? '#ece6e6' : '#2e2e2e',
-                        backgroundColor: isDark ? '#2e2e2e' : '#ece6e6',
-                    }}
-                    action={
-                        <IconButton
-                            aria-label="close"
-                            size="small"
-                            color="inherit"
-                            onClick={() => {
-                                setLocalStorageRecruitmentDismissalTime(Date.now().toString());
-                                setBannerVisibility(false);
-                            }}
-                        >
-                            <Close fontSize="inherit" />
-                        </IconButton>
-                    }
-                >
-                    Interested in web development?
-                    <br />
-                    <a href="https://forms.gle/v32Cx65vwhnmxGPv8" target="__blank" rel="noopener noreferrer">
-                        Join ICSSC and work on AntAlmanac and other projects!
-                    </a>
-                    <br />
-                    We have opportunities for experienced devs and those with zero experience!
-                </Alert>
-            ) : null}
-        </Box>
-    );
-};
-
-/* TODO: all this typecasting in the conditionals is pretty messy, but type guards don't really work in this context
- *  for reasons that are currently beyond me (probably something in the transpiling process that JS doesn't like).
- *  If you can find a way to make this cleaner, do it.
- */
-const SectionTableWrapped = (
-    index: number,
-    data: {
-        scheduleNames: string[];
-        courseData: (WebsocSchool | WebsocDepartment | AACourse)[];
-        larcData?: LarcAPIResponse | undefined;
-    }
-) => {
-    const { courseData, scheduleNames } = data;
-    const formData = RightPaneStore.getFormData();
-
-    let component;
-
-    if ((courseData[index] as WebsocSchool).departments !== undefined) {
-        const school = courseData[index] as WebsocSchool;
-        component = <SchoolDeptCard comment={school.schoolComment} type={'school'} name={school.schoolName} />;
-    } else if ((courseData[index] as WebsocDepartment).courses !== undefined) {
-        const dept = courseData[index] as WebsocDepartment;
-        component = <SchoolDeptCard name={`Department of ${dept.deptName}`} comment={dept.deptComment} type={'dept'} />;
-    } else if (formData.ge !== 'ANY') {
-        const course = courseData[index] as AACourse;
-        component = (
-            <GeDataFetchProvider
-                term={formData.term}
-                courseDetails={course}
-                allowHighlight={true}
-                scheduleNames={scheduleNames}
-                analyticsCategory={analyticsEnum.classSearch.title}
-            />
-        );
-    } else {
-        const course = courseData[index] as AACourse;
-        component = (
-            <SectionTableLazyWrapper
-                term={formData.term}
-                courseDetails={course}
-                allowHighlight={true}
-                scheduleNames={scheduleNames}
-                analyticsCategory={analyticsEnum.classSearch.title}
-            />
-        );
-    }
-
-    return <div>{component}</div>;
-};
-
-const LoadingMessage = () => {
-    const isDark = useThemeStore((store) => store.isDark);
-    return (
-        <Box sx={{ height: '100%', display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-            <img src={isDark ? darkModeLoadingGif : loadingGif} alt="Loading courses" />
-        </Box>
-    );
-};
-
-const ErrorMessage = () => {
-    const isDark = useThemeStore((store) => store.isDark);
-    return (
-        <Box sx={{ height: '100%', display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-            <img
-                src={isDark ? darkNoNothing : noNothing}
-                alt="No Results Found"
-                style={{ objectFit: 'contain', width: '80%', height: '80%' }}
-            />
-        </Box>
-    );
 };
 
 export default function CourseRenderPane(props: { id?: number }) {
@@ -314,34 +181,34 @@ export default function CourseRenderPane(props: { id?: number }) {
         };
     }, [setHoveredEvents]);
 
+    if (loading) {
+        return <LoadingMessage />;
+    }
+
+    if (error || courseData.length === 0) {
+        return <ErrorMessage />;
+    }
+
     return (
         <>
-            {loading ? (
-                <LoadingMessage />
-            ) : error || courseData.length === 0 ? (
-                <ErrorMessage />
-            ) : (
-                <>
-                    <RecruitmentBanner />
-                    <Box>
-                        <Box sx={{ height: '50px', marginBottom: '5px' }} />
-                        {courseData.map((_: WebsocSchool | WebsocDepartment | AACourse, index: number) => {
-                            let heightEstimate = 200;
-                            if ((courseData[index] as AACourse).sections !== undefined)
-                                heightEstimate = (courseData[index] as AACourse).sections.length * 60 + 20 + 40;
-                            return (
-                                <LazyLoad once key={index} overflow height={heightEstimate} offset={500}>
-                                    {SectionTableWrapped(index, {
-                                        courseData: courseData,
-                                        scheduleNames: scheduleNames,
-                                        larcData: larcResp,
-                                    })}
-                                </LazyLoad>
-                            );
-                        })}
-                    </Box>
-                </>
-            )}
+            <RecruitmentBanner />
+            <Box>
+                <Box sx={{ height: '50px', marginBottom: '5px' }} />
+                {courseData.map((_: WebsocSchool | WebsocDepartment | AACourse, index: number) => {
+                    let heightEstimate = 200;
+                    if ((courseData[index] as AACourse).sections !== undefined)
+                        heightEstimate = (courseData[index] as AACourse).sections.length * 60 + 20 + 40;
+                    return (
+                        <LazyLoad once key={index} overflow height={heightEstimate} offset={500}>
+                            {SectionTableWrapped(index, {
+                                courseData: courseData,
+                                scheduleNames: scheduleNames,
+                                larcData: larcResp,
+                            })}
+                        </LazyLoad>
+                    );
+                })}
+            </Box>
         </>
     );
 }
