+++ apps/antalmanac/src/components/RightPane/SectionTable/SchoolDeptCard.tsx
